Glossary component rename to GlossaryTable
